mix both solutions
