+++ packages/shared/src/index.ts
@@ -7,3 +7,4 @@ export * from "./component-utils";
 export * from "./crypto";
 export * from "./dropzone-utils";
 export * from "./tagged-errors";
+export * from "./content-disposition";
